backend/app/services/processor.py
"""Service layer for processing AI analysis."""

from typing import Dict, Any, Optional
import base64
import numpy as np
from io import BytesIO
from PIL import Image

# Import AI modules
import sys
from pathlib import Path
sys.path.insert(0, str(Path(__file__).parent.parent.parent.parent))

from ai_modules.face_module.face_detection import detect_faces
from ai_modules.gesture_module.gesture import recognize_gesture
from ai_modules.rppg_module.rppg import extract_rppg
from ai_modules.fusion.decision import make_decision


class AnalysisProcessor:
    """Main processor for analyzing flow state."""
    
    def __init__(self):
        """Initialize the processor."""
        self.face_detector = detect_faces
        self.gesture_recognizer = recognize_gesture
        self.rppg_extractor = extract_rppg
        self.decision_maker = make_decision
    
    def process_image(self, image_data: str) -> Dict[str, Any]:
        """
        Process image and perform all analyses.
        
        Args:
            image_data: Base64 encoded image
            
        Returns:
            Dictionary with analysis results
        """
        try:
            # Decode base64 image
            image = self._decode_image(image_data)
            
            # Run analyses
            face_data = self._analyze_face(image)
            gesture_data = self._analyze_gesture(image)
            rppg_data = self._analyze_rppg(image, face_data)
            
            # Fuse results
            flow_state = self._fuse_results(face_data, gesture_data, rppg_data)
            
            return {
                "success": True,
                "face_data": face_data,
                "gesture_data": gesture_data,
                "rppg_data": rppg_data,
                "flow_state": flow_state
            }
        except Exception as e:
            return {
                "success": False,
                "error": str(e)
            }
    
    def _decode_image(self, image_data: str) -> np.ndarray:
        """Decode base64 image to numpy array."""
        image_bytes = base64.b64decode(image_data)
        image = Image.open(BytesIO(image_bytes))
        return np.array(image)
    
    def _analyze_face(self, image: np.ndarray) -> Dict[str, Any]:
        """Analyze face in image."""
        # TODO: Call face detection module
        return {
            "detected": False,
            "faces": [],
            "confidence": 0.0
        }
    
    def _analyze_gesture(self, image: np.ndarray) -> Dict[str, Any]:
        """Analyze gestures in image."""
        # TODO: Call gesture recognition module
        return {
            "gesture": None,
            "confidence": 0.0
        }
    
    def _analyze_rppg(self, image: np.ndarray, face_data: Dict) -> Dict[str, Any]:
        """Analyze rPPG from face."""
        # TODO: Call rPPG extraction module
        return {
            "heart_rate": 0.0,
            "spo2": 0.0,
            "confidence": 0.0
        }
    
    def _fuse_results(self, face_data: Dict, gesture_data: Dict, rppg_data: Dict) -> Dict[str, Any]:
        """Fuse all results to determine flow state."""
        # TODO: Call fusion/decision module
        return {
            "flow_state": "unknown",
            "confidence": 0.0
        }
